Game Completable
Implemented an end of game screen which calculates the player's stars based on the amount of turns they took. Refactored Actors to only take 1 class for each type, instead of 1 for each colour using keyframes. Revised implementation of enemy/player turns to facilitate time between movements and more control in general. Implemented Fight, an object which controls conflicts between two actors of opposing types, allowing more control over how and when actors are destroyed. Minor tweaks to physics and collision handling. Fixed an issue that occurred when a Snake was blocked on both sides when attempting to move.