Implemented Scene Management
Implemented a scene manager which dictates what appears on the screen at any given time, stores the active scene and loads and unloads new scenes as they come into play. Each scene has 4 methods: Initialise() for initialising data members and events, LoadContent() for adding MovieClips to the scene, Update() which is called by SceneManager on every frame and UnloadContent() which is responsible for removing all necessary content from the screen to allow for the next scene to be loaded. The moment a scene's next:IScene variable is changed, the program will automatically call UnloadContent() and switch to that scene.

Added two scenes, StartMenuScene from Attila which is a basic menu with three buttons, and Level3Scene, inheriting from GameScene which contain all the logic of the original game up to and including the results screen.